pyproject.toml
# pyproject.toml

[project]
name = "pasteasfile"
version = "0.1.0"
description = "A system tray utility to paste clipboard text as a file."
requires-python = ">=3.8"
dependencies = [
    "pystray",
    "pillow",
    "pyperclip",
    "keyboard",
    "tkinterdnd2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]